LayoutTests/fast/sandbox/ios/sandbox-mach-lookup.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
<script>
description('Regression tests for mach lookup sandbox changes on iOS');

if (window.internals) {
    shouldBeFalse("internals.hasSandboxMachLookupAccessToXPCServiceName(\"com.matthewbenedict.CyberKit.WebContent\", \"com.apple.apple-extension-service\")");
    shouldBeFalse("internals.hasSandboxMachLookupAccessToXPCServiceName(\"com.matthewbenedict.CyberKit.WebContent\", \"com.apple.viewservice\")");
    shouldBeFalse("internals.hasSandboxMachLookupAccessToGlobalName(\"com.matthewbenedict.CyberKit.WebContent\", \"com.apple.TextInput\")");
    shouldBeFalse("internals.hasSandboxMachLookupAccessToGlobalName(\"com.matthewbenedict.CyberKit.WebContent\", \"com.apple.pluginkit.pkd\")");
    shouldBeFalse("internals.hasSandboxMachLookupAccessToGlobalName(\"com.matthewbenedict.CyberKit.WebContent\", \"com.apple.system.logger\")");
}
</script>
</head>
<body>
</body>
