LayoutTests/fast/sandbox/ios/sandbox-mach-lookup-expected.txt

Regression tests for mach lookup sandbox changes on iOS

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.hasSandboxMachLookupAccessToXPCServiceName("com.matthewbenedict.CyberKit.WebContent", "com.apple.apple-extension-service") is false
PASS internals.hasSandboxMachLookupAccessToXPCServiceName("com.matthewbenedict.CyberKit.WebContent", "com.apple.viewservice") is false
PASS internals.hasSandboxMachLookupAccessToGlobalName("com.matthewbenedict.CyberKit.WebContent", "com.apple.TextInput") is false
PASS internals.hasSandboxMachLookupAccessToGlobalName("com.matthewbenedict.CyberKit.WebContent", "com.apple.pluginkit.pkd") is false
PASS internals.hasSandboxMachLookupAccessToGlobalName("com.matthewbenedict.CyberKit.WebContent", "com.apple.system.logger") is false
